+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Current
 
 - Add a shortcut function `tradingstrategy.pair.generate_address_columns()` to generate `base_token_address` and `quote_token_address` columns
-- Some error message polish
+- Update API: `resolve_pairs_based_on_ticker()` supports `HumanReadableTradingPairDescription`
+- Update API: `resolve_pairs_based_on_ticker()` supports `HumanReadableTradingPairDescription`
+- Fix: Some error message polish
 
 # 0.13.18